Refined the teleportation system
Increased the teleportation distance from 5 to 50.
Adjusted the Raycast configuration for the Bezier curve to reach the reticle appropriately.
Turned off the left hand's ability to teleport.
Fixed the falling down bug that occurs during teleportation and game start.
Installed Unity's built in screen recorder to demonstrate the teleportation update.